deep_research/requirements.txt
gradio>=4.0
python-dotenv>=1.0
sendgrid>=6.0
pydantic>=2.0
openai>=1.0
requests>=2.0
